Security/GitHub dependabot (#173)
* chore(github): Add CODEOWNER

The reviewers field in the dependabot.yml file will be removed soon.
Please use the code owners file to specify reviewers for Dependabot PRs.
For more information, see this blog post.

https://github.blog/changelog/2025-04-29-dependabot-reviewers-configuration-option-being-replaced-by-code-owners/

* chore(deps): Change target branch to cooldown

This is to move maintainance tasks to cooldown

* chore(deps): Schedual for 6am Monday

Have this ready before work starts

* chore(deps): Configure commit message

* chore(deps): Remove hardcoded dependancies

* chore(deps): Update labels, Remove reviewers and assignees

The reviewers field in the dependabot.yml file will be removed soon.
Please use the code owners file to specify reviewers for Dependabot PRs.
For more information, see this blog post.
https://github.blog/changelog/2025-04-29-dependabot-reviewers-configuration-option-being-replaced-by-code-owners/

* fix(security): Remove extra whitespace in dependabot.yml